analytics/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.11
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
